build: rename settings module to definitions
